Revert "server: tests: remove db cleanup workaround"
This reverts commit 365d82755f650b7c94dc3e0075a868cd92143609.

+++ scripts/testsetup.js
@@ -49,7 +49,11 @@ const startFlood = () => {
 
 const closeProcesses = () => {
   floodProcess.on('close', () => {
-    fs.rmdirSync(temporaryRuntimeDirectory, {recursive: true});
+    if (process.env.CI !== 'true') {
+      // TODO: This leads to test flakiness caused by ENOENT error
+      // NeDB provides no method to close database connection
+      fs.rmdirSync(temporaryRuntimeDirectory, {recursive: true});
+    }
   });
 
   floodProcess.kill('SIGTERM');
