+++ server/.jest/auth.setup.js
@@ -11,5 +11,9 @@ process.argv.push('--auth', 'default');
 process.argv.push('--assets', 'false');
 
 afterAll(() => {
-  fs.rmdirSync(temporaryRuntimeDirectory, {recursive: true});
+  if (process.env.CI !== 'true') {
+    // TODO: This leads to test flakiness caused by ENOENT error
+    // NeDB provides no method to close database connection
+    fs.rmdirSync(temporaryRuntimeDirectory, {recursive: true});
+  }
 });
