+++ server/.jest/rtorrent.setup.js
@@ -31,6 +31,10 @@ process.argv.push('--assets', 'false');
 
 afterAll((done) => {
   process.kill(Number(fs.readFileSync(`${temporaryRuntimeDirectory}/rtorrent.pid`).toString()));
-  fs.rmdirSync(temporaryRuntimeDirectory, {recursive: true});
+  if (process.env.CI !== 'true') {
+    // TODO: This leads to test flakiness caused by ENOENT error
+    // NeDB provides no method to close database connection
+    fs.rmdirSync(temporaryRuntimeDirectory, {recursive: true});
+  }
   done();
 });
