+++ server/.jest/transmission.setup.js
@@ -43,7 +43,11 @@ process.argv.push('--assets', 'false');
 
 afterAll((done) => {
   transmissionProcess.on('close', () => {
-    fs.rmdirSync(temporaryRuntimeDirectory, {recursive: true});
+    if (process.env.CI !== 'true') {
+      // TODO: This leads to test flakiness caused by ENOENT error
+      // NeDB provides no method to close database connection
+      fs.rmdirSync(temporaryRuntimeDirectory, {recursive: true});
+    }
     done();
   });
